+++ packages/client/src/modules/users/components/index.tsx
@@ -1,0 +1,3 @@
+import { UsersDataGrid } from "./UserDataGrid";
+
+export { UsersDataGrid };
