+++ packages/client/src/modules/users/index.tsx
@@ -1,0 +1,1 @@
+export * from "./components";
